ES: renamed dead stuff elims
