+++ grin/src/Transformations/ExtendedSyntax/Optimising/DeadFunctionElimination.hs
@@ -1,83 +1,36 @@
-{-# LANGUAGE LambdaCase, RecordWildCards #-}
+{-# LANGUAGE LambdaCase, TupleSections, OverloadedStrings #-}
 module Transformations.ExtendedSyntax.Optimising.DeadFunctionElimination where
 
-import Data.Set (Set)
+
 import Data.Map (Map)
-import Data.Vector (Vector)
-
-import qualified Data.Set as Set
-import qualified Data.Map as Map
-import qualified Data.Vector as Vec
-
-import Data.List
-import Data.Maybe
-import Data.Monoid
-
-import qualified Data.Foldable
+import Data.Set (Set)
 import Data.Functor.Foldable as Foldable
 
-import Lens.Micro
-import Lens.Micro.Platform
+import qualified Data.Map as Map
+import qualified Data.Set as Set
+import qualified Data.Foldable
 
-import Control.Monad.Extra
-import Control.Monad.State
-import Control.Monad.Trans.Except
+import Text.Printf
 
 import Grin.ExtendedSyntax.Grin
-import Grin.ExtendedSyntax.Pretty
-import Grin.ExtendedSyntax.TypeEnv
-import Transformations.ExtendedSyntax.Util
-import AbstractInterpretation.ExtendedSyntax.LiveVariable.Result as LVA
 
+deadFunctionElimination :: Program -> Program
+deadFunctionElimination exp@(Program exts defs) = Program exts [def | def@(Def name _ _) <- defs, Set.member name liveDefs] where
+  defMap :: Map Name Def
+  defMap = Map.fromList [(name, def) | def@(Def name _ _) <- defs]
 
-type Trf = Except String
+  lookupDef :: Name -> Maybe Def
+  lookupDef name = Map.lookup name defMap
 
-runTrf :: Trf a -> Either String a
-runTrf = runExcept
+  liveDefs :: Set Name
+  liveDefs = fst $ until (\(live, visited) -> live == visited) visit (Set.singleton "grinMain", mempty)
 
-deadFunctionElimination :: LVAResult -> TypeEnv -> Exp -> Either String Exp
-deadFunctionElimination lvaResult tyEnv = runTrf .
-  (deleteDeadFunctions lvaResult >=> replaceDeadFunApps lvaResult tyEnv)
+  visit :: (Set Name, Set Name) -> (Set Name, Set Name)
+  visit (live, visited) = (mappend live seen, mappend visited toVisit) where
+    toVisit = Set.difference live visited
+    seen    = foldMap (maybe mempty (cata collect) . lookupDef) toVisit
 
-deleteDeadFunctions :: LVAResult ->  Exp -> Trf Exp
-deleteDeadFunctions lvaResult (Program exts defs) =
-  fmap (Program exts) $ filterM isFunDefLiveM defs where
-
-    isFunDefLiveM :: Exp -> Trf Bool
-    isFunDefLiveM (Def f _ _) = fmap not $ isRemovableM lvaResult f
-    isFunDefLiveM e = throwE $ "DFE: " ++ show (PP e) ++ " is not a function definition"
-
-
-replaceDeadFunApps :: LVAResult -> TypeEnv -> Exp -> Trf Exp
-replaceDeadFunApps lvaResult tyEnv = cataM alg where
-
-  alg :: ExpF Exp -> Trf Exp
-  alg = replaceAppWithUndefined lvaResult tyEnv . embed
-
-replaceAppWithUndefined :: LVAResult -> TypeEnv -> Exp -> Trf Exp
-replaceAppWithUndefined lvaResult TypeEnv{..} app@(SApp f _) = do
-  isRemovable <- isRemovableM lvaResult f
-  if isRemovable then do
-    (retTy,_) <- lookupExcept (notFoundInTyEnv f) f _function
-    pure $ SReturn $ Undefined (simplifyType retTy)
-  else
-    pure app
-  where notFoundInTyEnv f = "DFE: Function " ++ show (PP f) ++ " not found in type env"
-replaceAppWithUndefined _ _ e = pure e
-
-isRemovableM :: LVAResult -> Name -> Trf Bool
-isRemovableM lvaResult f = (&&) <$> isFunDeadM lvaResult f
-                                <*> hasNoSideEffectsM lvaResult f
-
-hasNoSideEffectsM :: LVAResult -> Name -> Trf Bool
-hasNoSideEffectsM LVAResult{..} f = fmap (not . _hasEffect)
-                                  . lookupExcept (noLiveness f) f
-                                  $ _functionEff
-
-isFunDeadM :: LVAResult -> Name -> Trf Bool
-isFunDeadM LVAResult{..} f = fmap isFunDead
-                           . lookupExcept (noLiveness f) f
-                           $ _functionLv
-
-noEffect   f = "DFE: Function " ++ show (PP f) ++ " not found in effect map"
-noLiveness f = "DFE: Function " ++ show (PP f) ++ " not found in liveness map"
+  collect :: ExpF (Set Name) -> Set Name
+  collect = \case
+    SAppF name _ | not (isExternalName exts name) -> Set.singleton name
+    exp -> Data.Foldable.fold exp
